refactor: separation of concern in the process endpoint
