+++ backend/src/app.py
@@ -1,8 +1,6 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-
-from src.models import GeneExpressionRequest
-from src.routers import gene_expression
+from routers import gene_expression
 
 app = FastAPI()
 
@@ -22,8 +20,3 @@ app.include_router(gene_expression.router, prefix="/api")
 @app.get("/")
 def read_root():
     return {"message": "Welcome to the Gene Expression Analysis API"}
-
-
-@app.post("/api/process")
-async def process_gene_expression(data: GeneExpressionRequest):
-    return {"message": "Request received", "data": data}
