+++ backend/src/routers/gene_expression.py
@@ -1,14 +1,9 @@
 from fastapi import APIRouter
-from pydantic import BaseModel
-from typing import List, Dict, Any
 import pandas as pd
-from src.database import get_db_connection
+from models import GeneExpressionData
+from database import get_db_connection
 
 router = APIRouter()
-
-
-class GeneExpressionData(BaseModel):
-    data: List[Dict[str, Any]]
 
 
 @router.post("/process")
